+++ examples/nodejs-logging/src/index.ts
@@ -2,7 +2,7 @@
 // SPDX-FileCopyrightText: 2022 Steadybit GmbH
 
 import { router as indexRouter } from './routes/index';
-import { router as loggingAttackRouter } from './routes/loggingAttack';
+import { router as loggingActionRouter } from './routes/logging';
 import express from 'express';
 import cors from 'cors';
 
@@ -13,10 +13,10 @@ app.use(express.json());
 
 app.use(cors());
 app.use(indexRouter);
-app.use(loggingAttackRouter);
+app.use(loggingActionRouter);
 
 app.listen(port, () => {
-	console.log(`Attack implementation listening on ${port}.`);
+	console.log(`Action implementation listening on ${port}.`);
 	console.log();
-	console.log(`Attack list can be accessed via GET http://127.0.0.1:${port}/actions`);
+	console.log(`Action list can be accessed via GET http://127.0.0.1:${port}/actions`);
 });