+++ examples/nodejs-logging/src/routes/index.ts
@@ -1,15 +1,16 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2022 Steadybit GmbH
 
-import { IndexResponse } from '@steadybit/action-api';
+import { ActionList } from '@steadybit/action-kit-api';
 import express from 'express';
 
 export const router = express.Router();
 
 router.get('/actions', (_, res) => {
-	const response: IndexResponse = {
-		attacks: [
+	const response: ActionList = {
+		actions: [
 			{
+				method: 'get',
 				path: '/actions/logging',
 			},
 		],